third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-line-and-column.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{
    InspectorTest.sendCommand("Runtime.enable", {});

    addConsoleMessagePromise("console.log(239)")
        .then(dumpMessage)
        .then(() => addConsoleMessagePromise("var l = console.log;\n  l(239)"))
        .then(dumpMessage)
        .then(() => InspectorTest.completeTest());

    function addConsoleMessagePromise(expression)
    {
        var cb;
        var p = new Promise((resolver) => cb = resolver);
        InspectorTest.eventHandler["Runtime.consoleAPICalled"] = (messageObject) => cb(messageObject);
        InspectorTest.sendCommand("Runtime.evaluate", { expression: expression });
        return p;
    }

    function dumpMessage(messageObject)
    {
        var msg = messageObject.params;
        delete msg.executionContextId;
        delete msg.args;
        delete msg.timestamp;
        for (var frame of msg.stackTrace.callFrames)
            frame.scriptId = 0;
        InspectorTest.logObject(msg);
    }
}

</script>
</head>
<body onload="runTest()">
</body>
</html>
