third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-line-and-column-expected.txt

CONSOLE MESSAGE: line 1: 239
CONSOLE MESSAGE: line 2: 239
{
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 8
                functionName : 
                lineNumber : 0
                scriptId : 0
                url : 
            }
        ]
    }
    type : log
}
{
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 2
                functionName : 
                lineNumber : 1
                scriptId : 0
                url : 
            }
        ]
    }
    type : log
}
